+++ client/src/cmdhf14a.c
@@ -2363,7 +2363,7 @@ int infoHF14A(bool verbose, bool do_nack_test, bool do_aid_search) {
         PrintAndLogEx(HINT, "Hint: try " _YELLOW_("`hf st info`"));
 
     if (isEMV)
-        PrintAndLogEx(HINT, "Hint: try " _YELLOW_("`emv search -s`"));
+        PrintAndLogEx(HINT, "Hint: try " _YELLOW_("`emv reader`"));
 
     if (isFUDAN) {
         PrintAndLogEx(HINT, "Hint: try " _YELLOW_("`hf fudan dump`"));